backend/pyproject.toml
[project]
name = "ask-my-agent"
version = "0.1.0"
description = "Digital-twin chatbot backend (FastAPI + ChromaDB + Anthropic/OpenAI)."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "slowapi>=0.1.9",
    "sse-starlette>=2.1",
    "chromadb>=0.5.3",
    "anthropic>=0.34",
    "openai>=1.40",
    "tiktoken>=0.7",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-telegram-bot[job-queue]>=21.0",
]

[project.optional-dependencies]
local-embed = ["sentence-transformers>=3.0"]
dev = ["pytest>=8.2", "pytest-asyncio>=0.23", "ruff>=0.5"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
